requirements.txt
streamlit==1.14.0
PyPDF2==3.0.0
pdf2image==1.16.3
openai==1.0.0
python-dotenv==0.21.0
Pillow==9.2.0
